requirements.txt
django >= 4.2.8, <= 5.0
django-ninja
psycopg2
djangorestframework-simplejwt